requirements.txt

pytest~=6.2.1
testcontainers~=3.4.0
kafka-python~=2.0.2
rootpath~=0.1.1